Move annotations.formula package
